26.NextWordPrediction-RNN,LSTM,GRU/requirements.txt
streamlit
tensorflow-cpu==2.15.0
numpy
pandas
pickle5
